requirements.txt
zeroc-ice
streamlit